Fix missing the endminusstart filter when set to zero
To be able to have a proper stepsize in the temporal controller, we set
the endminusstart value to 600 in case the user used a integration time of
zero. BUT while fixing the stepsize, this screws up the filter for the query.
So: we keep the chosen endminusstart, but when creating the (temporal) layer
we set it to a stepsize of 10 minutes (in case it is 0).